Add new API-level image thumbnailer; currently supports biographies/biography API calls, others pending
The API has traditionally provided an unsatisfactory method whereby the image filename is supplied, rather than a thumbnail.
Accordingly, this means that the client is responsible for converting this to an API and implementing thumbnailing.
This new implementation is a clean-room implementation, whereby the API output gives a URL to a thumbnail path, and the thumbnail will be created if not already present.